web-animations/animation/animation-target.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation target attribute test</title>
<meta name="assert" content="The element or pseudo-element being animated by this object">
<link rel="help" href="https://web-animations.github.io/web-animations-spec/bikeshed/Overview.html#dom-animation-target">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var target = createDiv(this);
    var animation = new Animation(target, null);
    animation.target = null;

    assert_equals(animation.target, null, 'Animation.target should be null');
}, 'Test setting target attribute to null value. Initial value of the attribute is not null');

test(function() {
    var animation = new Animation(null, null);
    var target = createDiv(this);
    animation.target = target;

    assert_equals(animation.target, target, 'Animation.target should be set by the attribute');
}, 'Test setting target attribute to not null value. Initial value of the attribute is null');

test(function() {
    var target1 = createDiv(this);
    var target2 = createDiv(this);
    var animation = new Animation(target1, null);
    animation.target = target2;

    assert_equals(animation.target, target2, 'Animation.target should be set by the attribute');
}, 'Test setting target attribute to not null value. Initial value of the attribute is not null');

test(function() {
    var element1 = createDiv(this);
    var element2 = createDiv(this);
    var target = element2.pseudo(':before');
    var animation = new Animation(target, null);
    animation.target = null;

    assert_equals(animation.target, null, 'Animation.target should be null');
}, 'Test setting target attribute to null value. Initial value of the attribute is ' +
    'not null PseudoElement');

test(function() {
    var animation = new Animation(null, null);
    var element1 = createDiv(this);
    var element2 = createDiv(this);
    var target = element2.pseudo(':before');
    animation.target = target;

    assert_equals(animation.target, target, 'Animation.target should be set by the attribute');
}, 'Test setting target attribute to not null PseudoElement. Initial value ' +
    'of the attribute is null');

test(function() {
    var target1 = createDiv(this);
    var element1 = createDiv(this);
    var element2 = createDiv(this);
    var target2 = element2.pseudo(':before');
    var animation = new Animation(target1, null);
    animation.target = target2;

    assert_equals(animation.target, target2, 'Animation.target should be set by the attribute');
}, 'Test setting target attribute to PseudoElement. Initial value of the attribute ' +
    'is Element');

test(function() {
    var element1 = createDiv(this);
    var element2 = createDiv(this);
    var target1 = element2.pseudo(':before');
    var target2 = createDiv(this);
    var animation = new Animation(target1, null);
    animation.target = target2;

    assert_equals(animation.target, target2, 'Animation.target should be set by the attribute');
}, 'Test setting target attribute to Element. Initial value of the attribute ' +
    'is PseudoElement');

test(function() {
    var element1 = createDiv(this);
    var element2 = createDiv(this);
    var target1 = element2.pseudo(':before');

    var element3 = createDiv(this);
    var element4 = createDiv(this);
    var target2 = element2.pseudo(':before');

    var animation = new Animation(target1, null);
    animation.target = target2;

    assert_equals(animation.target, target2, 'Animation.target should be set by the attribute');
}, 'Test setting target attribute to PseudoElement. Initial value of the attribute ' +
    'is another PseudoElement');
</script>
</body>
